180124 강의자료와 비교해보기
1. 길찾기 탐색 완료.
 강의 자료와 비교해보기

+++ Assets/01.Scripts/MainGame/Map/TileCell.cs
@@ -96,15 +96,15 @@ public class TileCell
 
     //pathfinding
     bool _isVisit = false;
-    int _distanceFromStart = 0;
-    int _distanceWeight = 1;
+    float _distanceFromStart = 0.0f;
+    float _distanceWeight = 1.0f;
 
     public void Visit() { _isVisit = true; }
     public bool IsVisit() { return _isVisit; }
     public void ResetVisit() { _isVisit = false; }
     public void ResetHeuristic() { _distanceFromStart = 0; }
-    public int GetDistanceFromStart() { return _distanceFromStart; }
-    public int GetDistanceWeight() { return _distanceWeight; }
+    public float GetDistanceFromStart() { return _distanceFromStart; }
+    public float GetDistanceWeight() { return _distanceWeight; }
 
     public void DrawColor()
     {
